fix: use int8 compute type for CPU compatibility
- Changed computeType from float16 to int8
- float16 requires GPU, int8 works on CPU
- Fixes 'Requested float16 compute type, but the target device or backend do not support efficient float16 computation' error

+++ src/workers/transcription-worker.ts
@@ -282,7 +282,7 @@ export class TranscriptionWorker {
         {
           model: 'medium',
           device: 'cpu',
-          computeType: 'float16',  // Changed from int8 to float16 (more stable)
+          computeType: 'int8',  // int8 for CPU compatibility (float16 requires GPU)
           beamSize: 5,
           vadFilter: false,  // Disable VAD for now (can enable later)
           wordTimestamps: false,
